requirements.txt
google-cloud-storage==1.31.2
pydub==0.24.1
google-cloud-speech==1.3.2
